# api/docs/admin/admin_swagger.yaml
basePath: /api
definitions:
  admin.AddressBookForm:
    properties:
      alias:
        type: string
      collection_id:
        type: integer
      forceAlwaysRelay:
        type: boolean
      hash:
        type: string
      hostname:
        type: string
      id:
        type: string
      loginName:
        type: string
      online:
        type: boolean
      password:
        type: string
      platform:
        type: string
      rdpPort:
        type: string
      rdpUsername:
        type: string
      row_id:
        type: integer
      sameServer:
        type: boolean
      tags:
        items:
          type: string
        type: array
      user_id:
        type: integer
      user_ids:
        items:
          type: integer
        type: array
      username:
        type: string
    required:
    - id
    type: object
...
  admin.ShareByWebClientForm:
    properties:
      expire:
        type: integer
      id:
        type: string
      password:
        type: string
      password_type:
        description: 只能是once,fixed
        enum:
        - once
        - fixed
        type: string
    required:
    - id
    - password
    - password_type
    type: object
...
  model.AddressBook:
    properties:
      alias:
        type: string
      collection:
        $ref: '#/definitions/model.AddressBookCollection'
      collection_id:
        type: integer
      created_at:
        type: string
      forceAlwaysRelay:
        type: boolean
      hash:
        type: string
      hostname:
        type: string
      id:
        type: string
      loginName:
        type: string
      online:
        type: boolean
      password:
        type: string
      platform:
        type: string
      rdpPort:
        type: string
      rdpUsername:
        type: string
      row_id:
        type: integer
      sameServer:
        type: boolean
      tags:
        items:
          type: string
        type: array
      updated_at:
        type: string
      user_id:
        type: integer
      username:
        type: string
    type: object
  model.AddressBookCollection:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    required:
    - name
    type: object
...
  model.AddressBookList:
    properties:
      list:
        items:
          $ref: '#/definitions/model.AddressBook'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: 接口
  title: 管理系统API
  version: "1.0"
paths:
  /admin/address_book/batchCreate:
    post:
      consumes:
      - application/json
      description: 批量创建地址簿
      parameters:
      - description: 地址簿信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.AddressBookForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AddressBook'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 批量创建地址簿
      tags:
      - 地址簿
  /admin/address_book/create:
    post:
      consumes:
      - application/json
      description: 创建地址簿
      parameters:
      - description: 地址簿信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.AddressBookForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AddressBook'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 创建地址簿
      tags:
      - 地址簿
  /admin/address_book/delete:
    post:
      consumes:
      - application/json
      description: 地址簿删除
      parameters:
      - description: 地址簿信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.AddressBookForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 地址簿删除
      tags:
      - 地址簿
  /admin/address_book/detail/{id}:
    get:
      consumes:
      - application/json
      description: 地址簿详情
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AddressBook'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 地址簿详情
      tags:
      - 地址簿
  /admin/address_book/list:
    get:
      consumes:
      - application/json
      description: 地址簿列表
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 页大小
        in: query
        name: page_size
        type: integer
      - description: 用户id
        in: query
        name: user_id
        type: integer
      - description: 是否是我的
        in: query
        name: is_my
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AddressBookList'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 地址簿列表
      tags:
      - 地址簿
  /admin/address_book/share:
    post:
      consumes:
      - application/json
      description: 地址簿分享
      parameters:
      - description: 地址簿信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.ShareByWebClientForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 地址簿分享
      tags:
      - 地址簿
  /admin/address_book/update:
    post:
      consumes:
      - application/json
      description: 地址簿编辑
      parameters:
      - description: 地址簿信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.AddressBookForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AddressBook'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 地址簿编辑
      tags:
      - 地址簿
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
  token:
    in: header
    name: api-token
    type: apiKey
swagger: "2.0"